Add app facade; move remaining types to shared folder
