+++ src/app/app.component.html
@@ -1,14 +1,14 @@
-<ng-container *ngIf="showModal$ | async">
+<ng-container *ngIf="facade.showModal$ | async">
 	<lcc-modal></lcc-modal>
 </ng-container>
 
-<ng-container *ngIf="showImageOverlay$ | async">
+<ng-container *ngIf="facade.showImageOverlay$ | async">
 	<lcc-image-overlay></lcc-image-overlay>
 </ng-container>
 
 <div class="lcc-app-container"
-		 (click)="onClickApp($event)">
-	<ng-container *ngIf="showAlert$ | async">
+		 (click)="facade.onClickApp($event)">
+	<ng-container *ngIf="facade.showAlert$ | async">
 		<lcc-alert></lcc-alert>
 	</ng-container>
 	<lcc-header></lcc-header>
@@ -18,7 +18,7 @@
 			<span *ngIf="isLoading"
 						class="loading-spinner">
 			</span>
-			<ng-container *ngIf="showToaster$ | async">
+			<ng-container *ngIf="facade.showToaster$ | async">
 				<lcc-toaster></lcc-toaster>
 			</ng-container>
 		</router-outlet>
