+++ src/app/core/auth/index.ts
@@ -5,7 +5,4 @@ export * as AuthActions from './store/auth.actions';
 export { AuthEffects } from './store/auth.effects';
 export * as AuthSelectors from './store/auth.selectors';
 export { reducer } from './store/auth.reducer';
-
-export { AuthState } from './types/auth.state';
-export { LoginRequest } from './types/login-request.model';
-export { SignUpRequest } from './types/sign-up-request.model';
+export { AuthState } from './store/auth.state';
