+++ src/app/core/nav/nav.facade.ts
@@ -9,14 +9,14 @@ import * as NavSelectors from './store/nav.selectors';
 
 @Injectable()
 export class NavFacade {
-  constructor(private readonly store: Store) {}
-
   user$ = this.store.select(AuthSelectors.user);
   isUserVerified$ = this.store.select(AuthSelectors.isUserVerified);
   isDropdownOpen$ = this.store.select(NavSelectors.isDropdownOpen);
 
-  // WIP: Need to find a good way to 'get member by ID' using their common UUID
+  // TODO: Need to find a good way to 'get member by ID' using their common UUID
   tempFirstName$ = of('Michal');
+
+  constructor(private readonly store: Store) {}
 
   onSelectHomeTab(): void {
     this.store.dispatch(NavActions.homeTabSelected());
