+++ src/app/core/nav/store/nav.selectors.ts
@@ -2,7 +2,7 @@ import { createFeatureSelector, createSelector } from '@ngrx/store';
 
 import { AppStoreFeatureTypes } from '@app/shared/types/app-store-features.model';
 
-import { NavState } from '../types/nav.state';
+import { NavState } from './nav.state';
 
 export const navFeatureSelector = createFeatureSelector<NavState>(
   AppStoreFeatureTypes.NAV
