+++ src/app/screens/schedule/index.ts
@@ -1,4 +1,2 @@
 export { ScheduleScreenComponent } from './schedule-screen.component';
 export { ScheduleScreenModule } from './schedule-screen.module';
-
-export { ClubEvent } from './types/club-event.model';
