+++ src/app/screens/sign-up/sign-up-screen.facade.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
 
-import { AuthActions, SignUpRequest } from '@app/core/auth';
+import { AuthActions } from '@app/core/auth';
+import { SignUpRequest } from '@app/shared/types';
 
 @Injectable()
 export class SignUpScreenFacade {
