+++ src/app/shared/components/alert/index.ts
@@ -4,7 +4,4 @@ export { AlertModule } from './alert.module';
 export * as AlertActions from './store/alert.actions';
 export * as AlertSelectors from './store/alert.selectors';
 export { reducer } from './store/alert.reducer';
-
-export { AlertAction } from './types/alert-action.model';
-export { Alert } from './types/alert.model';
-export { AlertState } from './types/alert.state';
+export { AlertState } from './store/alert.state';
