+++ src/app/shared/components/alert/store/alert.state.ts
@@ -1,0 +1,5 @@
+import { Alert } from '../../../types/alert.model';
+
+export interface AlertState {
+  alert?: Alert;
+}
