--- src/app/shared/components/alert/types/alert.model.ts
@@ -1,6 +1,0 @@
-import { AlertAction } from './alert-action.model';
-
-export interface Alert {
-  message: string;
-  action: AlertAction;
-}
