--- src/app/shared/components/alert/types/alert.state.ts
@@ -1,5 +1,0 @@
-import { Alert } from './alert.model';
-
-export interface AlertState {
-  alert?: Alert;
-}
