+++ src/app/shared/types/alert.model.ts
@@ -1,3 +1,8 @@
+export interface Alert {
+  message: string;
+  action: AlertAction;
+}
+
 export enum AlertAction {
   SEE_SCHEDULE = 'See schedule',
   REGISTER = 'Register',
