+++ src/app/shared/types/club-event.model.ts
