+++ src/app/shared/types/link.model.ts
@@ -1,4 +1,4 @@
-import { NavPathTypes } from '@app/core/nav';
+import { NavPathTypes } from './nav-paths.model';
 
 export interface Link {
   path: string | NavPathTypes;
