+++ src/app/shared/types/login-request.model.ts
