+++ src/app/shared/types/nav-paths.model.ts
