+++ src/app/shared/types/sign-up-response.model.ts
